+++ algo/src/main/java/com/hartwig/serve/ServeAlgo.java
@@ -19,10 +19,9 @@ import com.hartwig.serve.iclusion.classification.IclusionClassificationConfig;
 import com.hartwig.serve.iclusion.datamodel.IclusionTrial;
 import com.hartwig.serve.refgenome.RefGenomeManager;
 import com.hartwig.serve.refgenome.RefGenomeResource;
-import com.hartwig.serve.sources.ckb.CkbEvidenceExtractor;
+import com.hartwig.serve.sources.ckb.CkbExtractor;
 import com.hartwig.serve.sources.ckb.CkbExtractorFactory;
 import com.hartwig.serve.sources.ckb.CkbReader;
-import com.hartwig.serve.sources.ckb.CkbTrialExtractor;
 import com.hartwig.serve.sources.ckb.treatmentapproach.TreatmentApproachCurationEntry;
 import com.hartwig.serve.sources.ckb.treatmentapproach.TreatmentApproachCurationEntryKey;
 import com.hartwig.serve.sources.ckb.treatmentapproach.TreatmentApproachCurationFile;
@@ -155,10 +154,14 @@ public class ServeAlgo {
 
         TreatmentApproachCurator curator = new TreatmentApproachCurator(treatmentApproachMap);
 
-        CkbEvidenceExtractor extractor = CkbExtractorFactory.createEvidenceExtractor(config, refGenomeResource, curator);
+        CkbExtractor extractor = CkbExtractorFactory.createEvidenceExtractor(config, refGenomeResource, curator);
 
         LOGGER.info("Running CKB evidence knowledge extraction");
-        return extractor.extract(ckbEntries);
+        ExtractionResult result = extractor.extract(ckbEntries);
+
+        curator.reportUnusedCuratedEntries();
+
+        return result;
     }
 
     @NotNull
@@ -168,7 +171,7 @@ public class ServeAlgo {
         EventClassifierConfig config = CkbClassificationConfig.build();
         RefGenomeResource refGenomeResource = refGenomeManager.pickResourceForKnowledgebase(Knowledgebase.CKB_TRIAL);
 
-        CkbTrialExtractor extractor = CkbExtractorFactory.createTrialExtractor(config, refGenomeResource);
+        CkbExtractor extractor = CkbExtractorFactory.createTrialExtractor(config, refGenomeResource);
 
         LOGGER.info("Running CKB trial knowledge extraction");
         return extractor.extract(ckbEntries);
