+++ algo/src/main/java/com/hartwig/serve/sources/ckb/ActionableEntryFactory.java
@@ -7,7 +7,6 @@ import com.hartwig.serve.ckb.datamodel.CkbEntry;
 import org.jetbrains.annotations.NotNull;
 
 interface ActionableEntryFactory {
-
     @NotNull
-    Set<ActionableEntry> create(@NotNull CkbEntry entry, @NotNull String sourceGene, @NotNull String sourceEvent);
+    Set<ActionableEntry> create(@NotNull CkbEntry entry, @NotNull String sourceEvent, @NotNull String sourceGene);
 }
